% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation.R
\name{AIGENIE_checks}
\alias{AIGENIE_checks}
\title{AIGENIE Input Validation}
\usage{
AIGENIE_checks(
  item.attributes,
  openai.API,
  groq.API,
  custom,
  user.prompts,
  item.type.definitions,
  cleaning.fun,
  system.role,
  scale.title,
  sub.domain,
  model,
  item.examples,
  target.N,
  temperature,
  top.p,
  items.only,
  adaptive,
  EGA.model,
  EGA.algorithm,
  embedding.model,
  keep.org,
  plot,
  plot.stability,
  calc.final.stability,
  silently,
  ...
)
}
\arguments{
\item{item.attributes}{A named list or data frame containing item type labels and their corresponding attributes.}

\item{openai.API}{A required character string of your OpenAI API key.}

\item{groq.API}{(Required when using open-source models) A character string of your Groq API key.}

\item{custom}{Logical; indicates whether custom prompts and a custom cleaning function are used. Defaults to \code{FALSE}.}

\item{user.prompts}{(Required when \code{custom = TRUE}) A list of custom prompt strings for each item type.}

\item{item.type.definitions}{A named list or data frame providing definitions for each item type. Each definition should be a character string not exceeding 250 characters. This helps the language model understand the item types better. Definitions are included at the beginning of the prompts for their corresponding item types.}

\item{cleaning.fun}{(Required when \code{custom = TRUE}) A function provided by the user to clean and parse the model's output.}

\item{system.role}{An optional character string describing the language model's role.}

\item{scale.title}{An optional character string specifying the name of your inventory.}

\item{sub.domain}{An optional character string specifying the inventory's sub-domain or specialty.}

\item{model}{A character string specifying the model to use for item generation.}

\item{item.examples}{An optional character vector of example item strings.}

\item{target.N}{An integer or integer vector specifying the target number of items to generate.}

\item{temperature}{Numeric; controls the randomness of the model's output.}

\item{top.p}{Numeric; controls the diversity of the model's output.}

\item{items.only}{Logical; if \code{TRUE}, only items are generated without further analysis.}

\item{adaptive}{Logical; if \code{TRUE}, uses adaptive prompting to avoid generating redundant items.}

\item{EGA.model}{An optional character string specifying the EGA model to use (\code{"tmfg"} or \code{"glasso"}).}

\item{EGA.algorithm}{A character string specifying the clustering algorithm for EGA (default: \code{"walktrap"}).}

\item{embedding.model}{A character string specifying the OpenAI embedding model that should be used. The options are \code{"text-embedding-3-small"}, \code{"text-embedding-3-large"}, or \code{"text-embedding-ada-002"}. Defaults to \code{"text-embedding-3-small"}.}

\item{keep.org}{Logical; if \code{TRUE}, includes the original items in the returned results.}

\item{plot}{Logical; if \code{TRUE}, displays the network plots.}

\item{plot.stability}{Logical; Specifies whether to display the secondary network stability plots.}

\item{calc.final.stability}{Logical; defaults to \code{FALSE}. Specifies whether to compute the stability of the item pool before and after item reduction.}

\item{silently}{Logical; if \code{TRUE}, suppresses console output.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A list of validated and processed parameters ready for use in the \code{AIGENIE} function.
}
\description{
Validates and processes the input parameters for the \code{AIGENIE} function. This function ensures that all required parameters are provided, correctly formatted, and meet the necessary conditions for the AI-GENIE pipeline to run successfully. It handles both default and custom modes.
}
