% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{clean_EGA_output}
\alias{clean_EGA_output}
\title{Clean the EGA Output}
\usage{
clean_EGA_output(ega_obj, item_set)
}
\arguments{
\item{ega_obj}{The EGA object that is returned by the EGAnet function}

\item{item_set}{A data frame that contains the items and there attributes, types, and IDs.}
}
\value{
A list containing the EGA object and the item set with any of the items that did not belong to a community removed
}
\description{
Removes the items that were not assigned to a community silently should there be any.
}
