% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation_helpers.R
\name{clean_item_data}
\alias{clean_item_data}
\title{Clean Item Data}
\usage{
clean_item_data(item.data, string)
}
\arguments{
\item{item.data}{The item data to clean.}

\item{string}{A string used in error messages to specify the name of the data.}
}
\value{
A cleaned data frame with all columns as character type.
}
\description{
Trims whitespace and converts all columns in the item data to character type.
}
