% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{compute_ega_full_sample}
\alias{compute_ega_full_sample}
\title{Compute EGA on the full sample}
\usage{
compute_ega_full_sample(
  embedding,
  embedding_reduced,
  items,
  items_reduced,
  truth,
  truth_reduced,
  EGA.model,
  EGA.algorithm,
  title,
  calc.final.stability,
  silently
)
}
\arguments{
\item{embedding}{A matrix of embeddings for the all of the items.}

\item{embedding_reduced}{A matrix of embeddings for the items selected by AIGENIE.}

\item{items}{A data frame containing the item statements and types.}

\item{items_reduced}{A data frame containing the item statements and types for the items selected by AIGENIE.}

\item{truth}{A vector of the items' attribute assignment as generated by the model or coded in a provided data frame.}

\item{truth_reduced}{A vector of the items' attribute assignment as generated by the model or coded in a provided data frame for the items selected by AIGENIE.}

\item{EGA.model}{A character string specifying the EGA model to use (\code{"tmfg"} or \code{"glasso"}).}

\item{EGA.algorithm}{A character string specifying the EGA algorithm to be used.}

\item{title}{A character string of the title of the inventory}

\item{calc.final.stability}{Logical; a flag that determines whether stability should be calculated for the full sample.}

\item{silently}{Logical; if \code{TRUE}, suppresses console output.}

\item{...}{Additional arguments passed to underlying functions.}
}
\value{
A list containing the main results, final and initial EGA and bootEGA objects, embeddings, NMI values, and item counts before and after reduction.
}
\description{
Computes the Exploratory Graph Analysis (EGA) steps using the specified EGA model on the entire sample (i.e., not item-wise). Also determines the best embedding type to use and EGA model.
}
