% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{construct_item_examples_string}
\alias{construct_item_examples_string}
\title{Construct Formatted String of Example Items}
\usage{
construct_item_examples_string(item_examples, current_type)
}
\arguments{
\item{item_examples}{A validated data frame of item examples (output of \code{validate_item_examples_df()}).}

\item{current_type}{A character scalar indicating the type of items to include in the string.}
}
\value{
A character string with grouped and formatted item examples.
}
\description{
Given a validated item examples data frame, this function constructs
a string of well-formatted example items grouped by \code{attribute}, for a given \code{type}.
}
