% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation_helpers.R
\name{deduplicate_item_data}
\alias{deduplicate_item_data}
\title{Deduplicate Item Data}
\usage{
deduplicate_item_data(item.data, string)
}
\arguments{
\item{item.data}{The item data to deduplicate.}

\item{string}{A string used in error messages to specify the name of the data.}
}
\value{
A deduplicated data frame of item data.
}
\description{
Removes duplicate rows from the item data based on item statements and labels.
}
