% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation_helpers.R
\name{find_cols}
\alias{find_cols}
\title{Find Columns}
\usage{
find_cols(item.data, string)
}
\arguments{
\item{item.data}{The item data to process.}

\item{string}{A string used in error messages to specify the name of the data.}
}
\value{
A data frame with columns \code{statement} and \code{type}.
}
\description{
Attempts to determine which columns in the item data correspond to item statements and item types. Relies on the assumption that item statements are more unique than item types.
}
