% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{flatten_text}
\alias{flatten_text}
\title{Flatten Text}
\usage{
flatten_text(text)
}
\arguments{
\item{text}{A string that should be flattened}
}
\description{
Removes the punctuation from a string and converts it to lowercase.
}
