% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_output}
\alias{generate_output}
\title{Generate the Output ONLY for Users Testing a Custom Prompt}
\usage{
generate_output(
  openai.API,
  groq.API,
  user.prompts,
  N.runs,
  model,
  top.p,
  temperature,
  system.role,
  silently
)
}
\arguments{
\item{openai.API}{A string containing their OpenAI API key (or \code{NULL} if the user is using a Groq model)}

\item{groq.API}{A string containing their Groq API key (or \code{NULL} if the user is using an OpenAI model)}

\item{user.prompts}{A named list specifying the prompts that should be used to generate the output for each item type.}

\item{N.runs}{The number of times the code should generate output for each prompt.}

\item{model}{A string containing the LLM model to be used.}

\item{top.p}{The top-p setting of the model.}

\item{temperature}{The temperature setting of the model.}

\item{system.role}{A string containing the desired system role of the model}
}
\value{
A list of lists containing the output of the model for each item type across the runs.
}
\description{
Generates and returns a list of the text output that was generated by their prompts.
This can help them ensure there items are being written as expected. Also, this can
help users ensure their cleaning function works as expected.
}
