% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation.R
\name{GENIE_checks}
\alias{GENIE_checks}
\title{GENIE Input Validation}
\usage{
GENIE_checks(
  item.data,
  openai.API,
  EGA.model,
  EGA.algorithm,
  embedding.model,
  plot,
  plot.stability,
  calc.final.stability,
  silently
)
}
\arguments{
\item{item.data}{A required data frame containing your item statements and item type labels.}

\item{openai.API}{A required character string of your OpenAI API key.}

\item{EGA.model}{An optional character string specifying the EGA model to use (\code{"tmfg"} or \code{"glasso"}).}

\item{plot}{Logical; if \code{TRUE}, displays the network plots.}

\item{silently}{Logical; if \code{TRUE}, suppresses console output.}
}
\value{
A list containing:
\describe{
\item{\code{items}}{A cleaned and validated data frame of your item data.}
\item{\code{openai.API}}{Your validated OpenAI API key.}
}
}
\description{
Validates and processes the input parameters for the \code{GENIE} function. This function ensures that the provided item data is correctly formatted, contains no missing or duplicate values, and meets the requirements for a meaningful network analysis.
}
