% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_embeddings}
\alias{get_embeddings}
\title{Generates embeddings for the provided item statements using OpenAI's embedding model.}
\usage{
get_embeddings(items, dimensions = NULL, openai.key, ...)
}
\arguments{
\item{items}{A data frame containing the item statements. Must have a column named \code{statement}.}

\item{dimensions}{An optional integer specifying the number of dimensions for the embeddings. Defaults to \code{1536}.}

\item{openai.key}{A character string of your OpenAI API key.}
}
\value{
A matrix of embeddings where each column corresponds to an item statement.
}
\description{
Generates embeddings for the provided item statements using OpenAI's embedding model.
}
