% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{get_results}
\alias{get_results}
\title{Get Results}
\usage{
get_results(items, EGA.model, openai.key, item_type, keep.org, silently, ...)
}
\arguments{
\item{items}{A data frame containing the item statements and types.}

\item{EGA.model}{An optional character string specifying the EGA model to use (\code{"tmfg"} or \code{"glasso"}). If \code{NULL}, both models are evaluated, and the best one is selected.}

\item{openai.key}{A character string of your OpenAI API key.}

\item{item_type}{A character string of the current item type undergoing reduction}

\item{keep.org}{A logical that specifies whether or not the user wants to keep the original items}

\item{silently}{Logical; if \code{TRUE}, suppresses console output. Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed to underlying functions.}
}
\value{
A list containing the main results, EGA objects, bootEGA objects, embeddings, NMI values, and other analysis details.
}
\description{
Runs the AI-GENIE reduction analysis pipeline, including embeddings generation, redundancy removal using Unique Variable Analysis (UVA), instability removal, and EGA model selection based on Normalized Mutual Information (NMI).
}
