% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{llama_format}
\alias{llama_format}
\title{Llama Output Cleaning}
\usage{
llama_format(response, split_content)
}
\arguments{
\item{response}{A string that contains the output of the LLM model}

\item{split_content}{A vector of the item attributes stemmed}
}
\value{
A list of the characteristics and items extracted from the model output
}
\description{
Cleans the output text based on the expected typical output of the Llama model.
}
