% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{remove_instabilities}
\alias{remove_instabilities}
\title{Remove Instabilities}
\usage{
remove_instabilities(items, cut.off = 0.75, verbose, seed, model, ...)
}
\arguments{
\item{items}{A data matrix or data frame of items after redundancy removal.}

\item{cut.off}{Numeric; the stability cutoff value. Defaults to \code{0.75}.}

\item{...}{Additional arguments passed to the \code{\link[EGAnet]{bootEGA}} function.}
}
\value{
A data matrix or data frame with unstable items removed.
}
\description{
Performs Bootstrapped Exploratory Graph Analysis (bootEGA) to remove unstable items from the item pool based on a specified stability cutoff.
}
