% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation_helpers.R
\name{validate_and_extract_attributes}
\alias{validate_and_extract_attributes}
\title{Check the User provided dataframe and then extract the needed columns}
\usage{
validate_and_extract_attributes(df)
}
\arguments{
\item{df}{A data frame that the user provided containing the type, attribute, and statement}
}
\value{
A list of the validated item types and attributes
}
\description{
Checks that the provided data frame has the correct columns (type, attribute, and statement)
}
