% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation_helpers.R
\name{validate_apis}
\alias{validate_apis}
\title{Check the User provided API keys}
\usage{
validate_apis(openai.API, groq.API, model)
}
\arguments{
\item{openai.API}{A string that the user provided that should correspond to an OpenAI API string}

\item{groq.API}{A string that the user provided that should correspond to an Groq API string}

\item{model}{A string of the LLM model that user provided to generate items}
}
\value{
A cleaned list of the API string(s)
}
\description{
Checks that the provided at least one valid API string
}
