% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation_helpers.R
\name{validate_columns}
\alias{validate_columns}
\title{Validate Columns}
\usage{
validate_columns(item.data, string)
}
\arguments{
\item{item.data}{The item data to validate.}

\item{string}{A string used in error messages to specify the name of the data.}
}
\value{
No return value; the function stops with an error message if validation fails.
}
\description{
Ensures that the item data has exactly two columns and that at least one of them is of character type.
}
