% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation_helpers.R
\name{validate_custom}
\alias{validate_custom}
\title{Validate Custom Parameters}
\usage{
validate_custom(custom, item.attributes, user.prompts, cleaning.fun)
}
\arguments{
\item{custom}{Logical; a flag passed to the function to determine if the user added their own custom prompts}

\item{item.attributes}{A named list or data frame containing item type labels and their corresponding attributes. The list or data frame must have names or identifiers representing item types. Each element should be a character vector of attributes for that item type.}

\item{user.prompts}{A list of custom prompt strings for each item type.}

\item{cleaning.fun}{A function provided by the user to clean and parse the model's output.}
}
\value{
The validated \code{custom} flag.
}
\description{
Validates the parameters based on whether custom prompts are being used (\code{custom = TRUE}) or not. Ensures all required parameters are provided and appropriately sets others to \code{NULL} if necessary.
}
