% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation_helpers.R
\name{validate_EGA_algorithm}
\alias{validate_EGA_algorithm}
\title{Validate EGA Algorithm}
\usage{
validate_EGA_algorithm(EGA.algorithm)
}
\arguments{
\item{EGA.algorithm}{The input to be validated.}
}
\value{
A cleaned version of the EGA algorithm string
}
\description{
Checks if the provided EGA input is a string and is one of the accpted EGA algorithms
}
