% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation_helpers.R
\name{validate_embedding}
\alias{validate_embedding}
\title{Validate Embedding Model}
\usage{
validate_embedding(embedding)
}
\arguments{
\item{embedding}{A character string specifying the model to be used when embedding, one of  \code{"text-embedding-3-small"}, \code{"text-embedding-3-large"}, or \code{"text-embedding-ada-002"}.}
}
\value{
embedding string in lowercase; the function stops with an error message if validation fails.
}
\description{
Validates the \code{embedding.model} parameter to ensure it is either \code{"text-embedding-3-small"}, \code{"text-embedding-3-large"}, or \code{"text-embedding-ada-002"}.
}
