% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation_helpers.R
\name{validate_item_attributes}
\alias{validate_item_attributes}
\title{Validate Item Attributes}
\usage{
validate_item_attributes(item.attributes, items.only)
}
\arguments{
\item{item.attributes}{A named list or data frame where each element is a character vector of attributes corresponding to each item type.}

\item{items.only}{Logical; indicates if only items are to be generated without further processing.}
}
\value{
A list containing:
\describe{
\item{\code{labels}}{A character vector of item type labels.}
\item{\code{attribute}}{A named list of attributes for each item type.}
}
}
\description{
Ensures that the provided \code{item.attributes} is a named list or data frame with item type labels, that each item type has at least two attributes (unless \code{items.only = TRUE}), and that there are no duplicate stemmed item characteristics.
}
