% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation_helpers.R
\name{validate_item_data_type}
\alias{validate_item_data_type}
\title{Validate Item Data Type}
\usage{
validate_item_data_type(item.data, string)
}
\arguments{
\item{item.data}{The item data to validate.}

\item{string}{A string used in error messages to specify the name of the data.}
}
\value{
A data frame version of the item data.
}
\description{
Ensures that the provided item data is either a data frame or a matrix. Converts a matrix to a data frame if necessary.
}
