% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation_helpers.R
\name{validate_item_examples_p}
\alias{validate_item_examples_p}
\title{Validate Item Examples}
\usage{
validate_item_examples_p(item_examples, item_attributes, silently)
}
\arguments{
\item{item_examples}{A data frame containing columns: \code{type}, \code{attribute}, \code{item}, and \code{answer}.}

\item{item_attributes}{A named list where each name corresponds to an item type and each element is a character vector of subcategories.}

\item{silently}{A logical flag. If TRUE, the function will suppress messages regarding the example counts.}
}
\value{
A validated and deduplicated version of the \code{item.examples} data frame.
}
\description{
Validates the provided \code{item.examples} data frame. This function ensures that:
\enumerate{
\item \code{item.examples} is a data frame containing the required columns: \code{type}, \code{attribute}, \code{item}, and \code{answer}.
\item All columns contain only strings. If any column is a factor, it is converted to character.
\item There are no missing values in any of these columns.
\item The \code{type} values are one of the names in the provided \code{item.attributes} list.
\item The \code{attribute} values are one of the subcategories in the corresponding entry of \code{item.attributes}.
\item Duplicate rows are removed.
}
}
\details{
Additionally, the function checks the number of examples per item attribute:
\itemize{
\item If an attribute has no examples, a message is printed:
"For best results, provide at least one sample item per each item attribute. These attributes do not have any examples: \if{html}{\out{<list>}}"
\item If an attribute has more than 10 examples, a message is printed:
"These attributes have more than 10 examples: \if{html}{\out{<list>}}. Ensure you are using a model with a large context window."
}
}
