% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation_helpers.R
\name{validate_item_type_definitions}
\alias{validate_item_type_definitions}
\title{Check for Item Type Definitions in Parameters}
\usage{
validate_item_type_definitions(item.type.definitions, item.types)
}
\arguments{
\item{item.type.definitions}{A character list of the item type definitions}

\item{item.types}{A characater list of the item type labels}
}
\value{
No return value; the function stops with an error message if any \code{NA} values are found.
}
\description{
Checks provided parameters for any \code{NA} values and stops with an error if any are found.
}
