% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation_helpers.R
\name{validate_no_duplicate_items}
\alias{validate_no_duplicate_items}
\title{Validate No Duplicate Items}
\usage{
validate_no_duplicate_items(item.data, string)
}
\arguments{
\item{item.data}{The item data to validate.}

\item{string}{A string used in error messages to specify the name of the data.}
}
\value{
No return value; the function stops with an error message if duplicates are found.
}
\description{
Checks for duplicate item statements across different labels and stops with an error if any are found.
}
