% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation.R
\name{validate_promt_inputs}
\alias{validate_promt_inputs}
\title{GENIE Input Validation}
\usage{
validate_promt_inputs(
  openai.API,
  groq.API,
  user.prompts,
  N.runs,
  model,
  top.p,
  temperature,
  system.role,
  silently
)
}
\arguments{
\item{openai.API}{Optional (Required if using a GPT model). A character string containing your OpenAI API key.}

\item{groq.API}{Optional (Required if using a Groq model). A character string containing your Groq API key.}

\item{user.prompts}{A named list of custom prompt strings for each item type.}

\item{N.runs}{An integer specifying the number of times to run the prompt for preview; defaults to \code{3}.}

\item{model}{A character string specifying the language model to use. Options include \code{"gpt3.5"}, \code{"gpt4o"},
\code{"llama3"}, \code{"mixtral"}, \code{"deepseek"}, or \code{"gemma2"}. Defaults to \code{"gpt3.5"}.}

\item{top.p}{Numeric; defaults to \code{1}. Sets the top-p sampling parameter for the language model.}

\item{temperature}{Numeric; defaults to \code{1}. Controls the randomness of the model's output (valid range: 0–2).}

\item{system.role}{Optional. A character string defining the role of the language model (e.g., "an expert methodologist").}

\item{silently}{Logical; defaults to \code{FALSE}. If \code{TRUE}, suppresses console output.}
}
\value{
A named list containing all of the cleaned and validated user parameters
}
\description{
Validates and processes the input parameters for the \code{custom_prompt} function. This function ensures that the provided item data is correctly formatted, contains no missing or duplicate values, and meets the requirements.
}
