% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation_helpers.R
\name{validate_string_vector}
\alias{validate_string_vector}
\title{Validate String Vector}
\usage{
validate_string_vector(x)
}
\arguments{
\item{x}{The input to be validated.}
}
\value{
A logical value: \code{TRUE} if \code{x} is a vector of strings, \code{FALSE} otherwise.
}
\description{
Checks if the provided input is a vector where all elements are strings (character type).
}
