% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input_validation_helpers.R
\name{validate_temperature_top.p}
\alias{validate_temperature_top.p}
\title{Validate Temperature and Top P}
\usage{
validate_temperature_top.p(temperature, top.p)
}
\arguments{
\item{temperature}{A numeric value between 0 and 2 used to set the temperature of the language model.}

\item{top.p}{A numeric value between 0 and 1 used to set the top-p sampling of the language model.}
}
\value{
A list containing validated \code{temperature} and \code{top.p} values.
}
\description{
Validates the \code{temperature} and \code{top.p} parameters to ensure they are numeric and within the valid range.
}
