% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{create.prompts}
\alias{create.prompts}
\title{Create Prompts}
\usage{
create.prompts(
  item.attributes,
  item.type.definitions,
  scale.title,
  sub.domain,
  item.examples,
  system.role
)
}
\arguments{
\item{item.attributes}{A list where each element is a vector of attributes corresponding to each item type.}

\item{item.type.definitions}{An optional named list or data frame providing definitions for each item type. Each definition should be a character string not exceeding 250 characters. This helps the language model understand the item types better. Definitions are included at the beginning of the prompts for their corresponding item types.}

\item{scale.title}{An optional character string specifying the name of your inventory.}

\item{sub.domain}{An optional character string specifying the inventory's sub-domain or specialty.}

\item{item.examples}{An optional character vector of example item strings.}

\item{system.role}{An optional character string describing the language model's role.}
}
\value{
A list containing:
\describe{
\item{\code{user.prompts}}{A list of user prompts for each item type to instruct the language model.}
\item{\code{system.role}}{A character string for the system role prompt.}
}
}
\description{
Generates user prompts and a system role prompt for item generation based on provided item types, attributes, scale title, sub-domain, and optional item examples.
}
