% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{flatten_text}
\alias{flatten_text}
\title{Flatten Text}
\usage{
flatten_text(text)
}
\arguments{
\item{text}{A character string or vector to be flattened.}
}
\value{
A character string or vector with text converted to lowercase and punctuation removed.
}
\description{
Processes text by converting it to lowercase and removing punctuation.
}
