% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_networks}
\alias{plot_networks}
\title{Plot Networks Before and After AI-GENIE Reduction}
\usage{
plot_networks(p1, p2, nmi2, nmi1, scale.title)
}
\arguments{
\item{p1}{An EGA object representing the initial network structure.}

\item{p2}{An EGA object representing the final network structure after AI-GENIE reduction.}

\item{nmi2}{Numeric; the Normalized Mutual Information (NMI) of the final item pool.}

\item{nmi1}{Numeric; the NMI of the original item pool.}

\item{scale.title}{A character string specifying the title for the combined plot.}
}
\value{
A combined plot object displaying the initial and final networks with captions and annotations.
}
\description{
Creates a combined plot of the initial and final network structures before and after AI-GENIE reduction, including NMI values and a change subtitle.
}
