% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run AI-GENIE Pipeline}
\usage{
run_pipeline(
  items,
  openai.key,
  title = "Networks Before and After AI-Genie",
  EGA_model = NULL,
  keep.org = FALSE,
  plot = TRUE,
  plot.stability = FALSE,
  silently = FALSE,
  ...
)
}
\arguments{
\item{items}{A data frame containing the item statements and types.}

\item{openai.key}{A character string of your OpenAI API key.}

\item{title}{An optional character string specifying the title for the network plots. Defaults to \code{"Networks Before and After AI-Genie"}.}

\item{EGA_model}{An optional character string specifying the EGA model to use (\code{"tmfg"} or \code{"glasso"}). If \code{NULL}, both models are evaluated, and the best one is selected.}

\item{keep.org}{Logical; if \code{TRUE}, includes the original items in the returned results. Defaults to \code{FALSE}.}

\item{plot}{Logical; if \code{TRUE}, displays the network plots. Defaults to \code{TRUE}.}

\item{plot.stability}{Logical; Specifies whether to display the secondary network stability plots. Defaults to \code{FALSE}.}

\item{silently}{Logical; if \code{TRUE}, suppresses console output. Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed to underlying functions.}
}
\value{
A list containing:
\describe{
\item{\code{main_result}}{A data frame of the item pool after AI-GENIE reduction. The data frame has the columns \code{ID}, \code{type}, \code{statement}, and \code{EGA_communities}.}
\item{\code{final_ega_obj}}{The final EGA object after reduction.}
\item{\code{final_bootega_obj}}{The final bootEGA object after reduction.}
\item{\code{initial_ega_obj}}{The initial EGA object with the entire item pool.}
\item{\code{initial_bootega_obj}}{The initial bootEGA object generated from redundancy-reduced data.}
\item{\code{embeddings}}{The embeddings generated for the items.}
\item{\code{embedding_type}}{The type of embeddings used ("sparse" or "full").}
\item{\code{selected_model}}{The EGA model used throughout the pipeline.}
\item{\code{nmi}}{The Normalized Mutual Information (NMI) of the final item pool.}
\item{\code{start_nmi}}{The NMI of the original item pool.}
\item{\code{start_N}}{The starting sample size (number of items).}
\item{\code{final_N}}{The final sample size after reduction.}
\item{\code{original_items (optional)}}{(ONLY returns if \code{keep.org} is \code{TRUE}) The original sample generated.}
}
}
\description{
Executes the AI-GENIE reduction analysis pipeline on the provided items, including embeddings generation, redundancy removal, instability removal, and network plotting.
}
